pyproject.toml
[project]
name = "dataflow"
version = "0.1.0"
requires-python = ">=3.13"
description = "Config-driven data pipeline platform"
dependencies = [
    "pydantic>=2.0,<3.0",
    "pandas>=2.0",
    "duckdb>=0.10",
    "apscheduler>=3.10",
    "pymysql>=1.1",
    "psycopg2-binary>=2.9",
    "boto3>=1.34",
    "requests>=2.31",
    "click>=8.1",
    "python-dotenv>=1.0",
    "pyarrow>=15.0",
    "pytz>=2024.1",
    "fastapi[standard]>=0.135.2",
]

[project.scripts]
dataflow = "main:cli"

[tool.setuptools]
packages = ["dataflow"]
py-modules = ["main"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
